pyproject.toml
[project]
name = "multi-survey-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=1.0.22",
    "livekit-plugins-noise-cancellation>=0.2.4",
    "livekit-plugins-openai>=1.0.22",
    "pandas>=2.2.3",
    "python-dotenv~=1.0",
    "pyyaml>=6.0.2",
]
